refactor: consolidate types and add environment-based Member Portal switching
