+++ src/env.ts
@@ -25,6 +25,12 @@ const envSchema = z.object({
     .default("false")
     .transform((v) => v === "true"),
 
+  // Member Portal API integration toggle
+  USE_MOCK_MEMBER_API: z
+    .string()
+    .default("true")
+    .transform((v) => v === "true"),
+
   // Unsubscribe token signing secret (256-bit minimum)
   UNSUBSCRIBE_SECRET: z.string().min(32),
