+++ src/lib/action-types.ts
@@ -1,0 +1,5 @@
+export interface ActionResult<T = void> {
+  success: boolean;
+  data?: T;
+  error?: string;
+}
